+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/WiringHarness.java
@@ -2,12 +2,8 @@ package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.hardware.rev.Rev2mDistanceSensor;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
-import com.qualcomm.robotcore.hardware.LED;
 import com.qualcomm.robotcore.hardware.Servo;
 
 import com.qualcomm.robotcore.hardware.DistanceSensor;
@@ -31,7 +27,6 @@ import org.openftc.easyopencv.OpenCvPipeline;
 
 import android.app.Activity;
 import android.graphics.Color;
-import android.graphics.ColorSpace;
 import android.view.View;
 
 @Autonomous
@@ -391,13 +386,11 @@ public class WiringHarness extends RobotLinearOpMode{
 }
 
 
-class SkystoneDeterminationPipelineRedFar extends OpenCvPipeline
-{
+class SkystoneDeterminationPipelineRedFar extends OpenCvPipeline {
     /*
      * An enum to define the skystone position
      */
-    public enum SkystonePosition
-    {
+    public enum SkystonePosition {
         LEFT,
         CENTER,
         RIGHT
@@ -412,9 +405,9 @@ class SkystoneDeterminationPipelineRedFar extends OpenCvPipeline
     /*
      * The core values which define the location and size of the sample regions
      */
-    static final Point REGION1_TOPLEFT_ANCHOR_POINT = new Point(0,60);
-    static final Point REGION2_TOPLEFT_ANCHOR_POINT = new Point(107,60);
-    static final Point REGION3_TOPLEFT_ANCHOR_POINT = new Point(213,60);
+    static final Point REGION1_TOPLEFT_ANCHOR_POINT = new Point(0, 60);
+    static final Point REGION2_TOPLEFT_ANCHOR_POINT = new Point(107, 60);
+    static final Point REGION3_TOPLEFT_ANCHOR_POINT = new Point(213, 60);
     static final int REGION_WIDTH = 106;
     static final int REGION_HEIGHT = 160;
 
@@ -466,20 +459,19 @@ class SkystoneDeterminationPipelineRedFar extends OpenCvPipeline
 
     // Volatile since accessed by OpMode thread w/o synchronization
     private volatile SkystonePosition position = SkystonePosition.RIGHT;
+    private volatile int colorLookingAt = -1;
 
     /*
      * This function takes the RGB frame, converts to YCrCb,
      * and extracts the Cb channel to the 'Cb' variable
      */
-    void inputToCb(Mat input)
-    {
+    void inputToCb(Mat input) {
         Imgproc.cvtColor(input, YCrCb, Imgproc.COLOR_RGB2YCrCb);
         Core.extractChannel(YCrCb, Cr, 1);
     }
 
     @Override
-    public void init(Mat firstFrame)
-    {
+    public void init(Mat firstFrame) {
 
         /*
          * We need to call this in order to make sure the 'Cb'
@@ -503,8 +495,11 @@ class SkystoneDeterminationPipelineRedFar extends OpenCvPipeline
     }
 
     @Override
-    public Mat processFrame(Mat input)
-    {
+    public Mat processFrame(Mat input) {
+
+        /* This is the code for color detection */
+        colorLookingAt = ColorSense.get_color_of_brick(input);
+
         /*
          * Overview of what we're doing:
          *
@@ -600,7 +595,7 @@ class SkystoneDeterminationPipelineRedFar extends OpenCvPipeline
          * Now that we found the max, we actually need to go and
          * figure out which sample region that value was from
          */
-        if(min == avg1) // Was it from region 1?
+        if (min == avg1) // Was it from region 1?
         {
             position = SkystonePosition.LEFT; // Record our analysis
 
@@ -614,8 +609,7 @@ class SkystoneDeterminationPipelineRedFar extends OpenCvPipeline
                     region1_pointB, // Second point which defines the rectangle
                     GREEN, // The color the rectangle is drawn in
                     -1); // Negative thickness means solid fill
-        }
-        else if(min == avg2) // Was it from region 2?
+        } else if (min == avg2) // Was it from region 2?
         {
             position = SkystonePosition.CENTER; // Record our analysis
 
@@ -629,8 +623,7 @@ class SkystoneDeterminationPipelineRedFar extends OpenCvPipeline
                     region2_pointB, // Second point which defines the rectangle
                     GREEN, // The color the rectangle is drawn in
                     -1); // Negative thickness means solid fill
-        }
-        else if(min == avg3) // Was it from region 3?
+        } else if (min == avg3) // Was it from region 3?
         {
             position = SkystonePosition.RIGHT; // Record our analysis
 
@@ -657,10 +650,224 @@ class SkystoneDeterminationPipelineRedFar extends OpenCvPipeline
     /*
      * Call this from the OpMode thread to obtain the latest analysis
      */
-    public SkystonePosition getAnalysis()
-    {
+    public SkystonePosition getAnalysis() {
         return position;
     }
+    public int getColor() {return colorLookingAt;}
 }
+class ColorSense {
+    static final int HSV_MAX = 180;
+    static final int COLOR_RECOGNITION_THRESHOLD = 80;
+    static final int[][] RANGES = {{170, 10}, {10, 60}, {105, 135}};
+
+    /*
+     * 0: Red
+     * 1: Yellow
+     * 2: Blue
+     */
+
+    // HSV values are 1 to 180, so if you have an interval like (170, 10)
+    // you have to change it to (170, 180), (0, 10)
+    // otherwise just return the interval
+    public static int[][] get_ivals(int[] colorRange){
+        int[][] ivals = new int[2][2];
+        if (colorRange[0] > colorRange[1]){
+            // case like (170, 10)
+            ivals[0][0] = colorRange[0];
+            ivals[0][1] = HSV_MAX;
+
+            ivals[1][0] = 0;
+            ivals[1][1] = colorRange[1];
+        }
+        else {
+            // normal case
+            ivals[0][0] = colorRange[0];
+            ivals[0][1] = colorRange[1];
+        }
+        return ivals;
+    }
+
+    private static int get_color(double[] HSV){
+        /*
+         * I intend HSV[0] to be HUE, [1] to be SAT, and [2] to be VAL.
+         *
+         * If returned 0, then blue
+         * If returned 1, then red
+         * If returned 2, then yellow
+         * If none found, return 3
+         */
+
+        // too dark / too unsaturated
+        if (!(50 <= HSV[1] || HSV[2] <= 255)){
+            return 3;
+        }
+        // iterate through all colors
+        for (int color = 0; color < 3; color++){
+            int[][] ival = get_ivals(RANGES[color]);
+
+            for (int[] ints : ival) {
+                if (ints[0] != 0 || ints[1] != 0) {
+                    // inside color range
+                    if (ints[0] <= HSV[0] && HSV[0] <= ints[1]) {
+                        return color;
+                    }
+                }
+            }
+        }
+        // none were in range
+        return 3;
+    }
+
+    private static double max(double[] values){
+        // Given that all values
+        double maxVal = Double.MIN_VALUE;
+
+        for (double val : values) {
+            if (val > maxVal){
+                maxVal = val;
+            }
+        }
+        return maxVal;
+    }
+
+    private static double min(double[] values){
+        // Given that all values
+        double minVal = Double.MAX_VALUE;
+
+        for (double val : values) {
+            if (val < minVal){
+                minVal = val;
+            }
+        }
+        return minVal;
+    }
+
+    // public static int[] convert_rgb_to_hsv(int[] RGB){
+    //     /* In this function, we assume that RGB is an int array with ranges 0-255,
+    //     * where index 0 is red, 1 is green, 2 is blue.
+    //     * We first convert the RGB value to a percentage between 0 and 1.
+    //     */
+    //     double[] RGBPercents = {RGB[0] / 255.0, RGB[1] / 255.0, RGB[2] / 255.0};
+    //     double CMAX = max(RGBPercents);
+    //     double CMIN = min(RGBPercents);
+
+    //     double delta = CMAX - CMIN;
+
+    //     int[] HSV = new int[3];
+
+    //     // This first chunk of code is to get the Hue value.
+    //     if (delta == 0){
+    //         HSV[0] = 0;
+    //     }
+    //     else if (CMAX == RGBPercents[0]){
+    //         HSV[0] = Math.toIntExact(Math.round(60 * (((RGBPercents[1] - RGBPercents[2]) / delta) % 6)));
+    //     }
+    //     else if (CMAX == RGBPercents[1]){
+    //         HSV[0] = Math.toIntExact(Math.round(60 * ((RGBPercents[2] - RGBPercents[0]) / delta + 2)));
+    //     }
+    //     else if (CMAX == RGBPercents[2]){
+    //         HSV[0] = Math.toIntExact(Math.round(60 * ((RGBPercents[0] - RGBPercents[1]) / delta + 4)));
+    //     }
+
+    //     // This code gets the Saturation.
+    //     if (CMAX == 0){
+    //         HSV[1] = 0;
+    //     }
+    //     else {
+    //         HSV[1] = Math.toIntExact(Math.round(delta / CMAX));
+    //     }
+
+    //     // Value is simply CMAX
+    //     HSV[2] = Math.toIntExact(Math.round(CMAX));
+
+    //     return HSV;
+    // }
+
+    public static int get_color_of_brick(Pixel[] frame){
+        // get count of each pixel
+        int[] count = new int[4];
+        /*
+         * Edwin uses a dictionary here. Java has no such dictionary as I'm aware
+         * of, so I'm going to instead omit the "key" part of the array.
+         * count[0] is for blue, count[1] is for red, count[2] is for yellow
+         * similarly to how the get_color function output is coded.
+         * count[4] is for no color found
+         * If returned 0, then blue
+         * If returned 1, then red
+         * If returned 2, then yellow
+         * If none found, return 3
+         */
+        double total_pixels = frame.length;
+        for (Pixel pixel : frame){
+            count[get_color(pixel.getHSV())]++;
+        }
+
+        // if a color passes the threshold we are confident in the fact that there is a correct color brick underneath us
+
+        for (int color = 0; color < 3; color++){
+            if (count[color] / total_pixels > COLOR_RECOGNITION_THRESHOLD){
+                return color;
+            }
+        }
+        /* This return statement might be unnecessary.
+         * Because, in the previous for loop, if the majority of pixels have no color
+         * found, then it will pass the recognition threshold and return no color.
+         * Leaving it in though, because I am a faithful translator
+         */
+        return 3;
+    }
+
+    public static int get_color_of_brick(Mat input){
+        return get_color_of_brick(convertMatToPixel(input));
+    }
+
+    private static Pixel[] convertMatToPixel(Mat mat){
+        int rows = mat.rows();
+        int cols = mat.cols();
+
+        Mat HSVs = new Mat();
+
+        Imgproc.cvtColor(mat, HSVs, Imgproc.COLOR_RGB2HSV);
+
+        Pixel[] pixels = new Pixel[rows * cols];
+
+        int counter = 0;
+        for (int row = 0; row < rows; row++){
+            for (int col = 0; col < cols; col++){
+                /* the mat is going to be called firstFrame and is in init.
+                 * Assume type inside of HSVs array is double[] */
+                pixels[counter++] = new Pixel(mat.get(row, col));
+            }
+        }
+        return pixels;
+    }
+
+}
+class Pixel {
+    private double[] HSV = new double[3];
+
+    public Pixel(){
+
+    }
+    public Pixel(double hue, double sat, double val){
+        HSV[0] = hue;
+        HSV[1] = sat;
+        HSV[2] = val;
+    }
+
+    public Pixel(double[] HSV){
+        if (HSV.length == 3){
+            this.HSV = HSV;
+        }
+        else{
+            this.HSV[0] = HSV[0];
+            this.HSV[1] = HSV[1];
+            this.HSV[2] = HSV[2];
+        }
+    }
 
 
+    public double[] getHSV(){
+        return HSV;
+    }
+}
